server/pyproject.toml
[project]
name = "cellim-viewer-server"
version = "0.0.0"
description = "Server for CELLIM Viewer."
readme = "README.md"
requires-python = "==3.12.9"
dependencies = [
    "fastapi==0.115.11",
    "pydantic-settings==2.8.1",
    "psycopg==3.2.6",
    "sqlalchemy==2.0.39",
    "uvicorn==0.34.0",
]

[dependency-groups]
dev = [
    "mypy==1.15.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    # isort
    "I",
]
